+++ vitest.config.js
@@ -2,6 +2,10 @@ import { cloudflareTest } from "@cloudflare/vitest-pool-workers";
 import { defineConfig } from "vitest/config";
 
 export default defineConfig({
+  test: {
+    include: ["test/**/*.test.js"],
+    exclude: ["test/contracts/**"],
+  },
   plugins: [
     cloudflareTest({
       wrangler: {